Use base handler for user invitation job
